moved Type into language
