+++ src/inspection/type/common.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { Type } from "../../type";
+import { Type } from "../../language";
 
 export type TypeById = Map<number, Type.TType>;
