+++ src/inspection/type/inspectType/inspectTypeLiteralExpression.ts
@@ -2,9 +2,8 @@
 // Licensed under the MIT license.
 
 import { Assert } from "../../../common";
-import { Ast } from "../../../language";
+import { Ast, Type, TypeUtils } from "../../../language";
 import { TXorNode, XorNodeKind, XorNodeUtils } from "../../../parser";
-import { Type, TypeUtils } from "../../../type";
 
 export function inspectTypeLiteralExpression(xorNode: TXorNode): Type.TType {
     XorNodeUtils.assertAstNodeKind(xorNode, Ast.NodeKind.LiteralExpression);
