+++ src/inspection/type/inspectType/inspectTypeParameter.ts
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { Ast } from "../../../language";
+import { Ast, Type } from "../../../language";
 import { NodeIdMapUtils, TXorNode, XorNodeUtils } from "../../../parser";
-import { Type } from "../../../type";
 import { inspectTypeFromChildAttributeIndex, TypeInspectionState } from "./common";
 
 export function inspectTypeParameter(state: TypeInspectionState, xorNode: TXorNode): Type.TType {
