+++ src/inspection/type/inspectType/inspectTypeRecord.ts
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { Ast } from "../../../language";
+import { Ast, Type } from "../../../language";
 import { NodeIdMapIterator, TXorNode, XorNodeUtils } from "../../../parser";
-import { Type } from "../../../type";
 import { inspectXorNode, TypeInspectionState } from "./common";
 
 export function inspectTypeRecord(state: TypeInspectionState, xorNode: TXorNode): Type.DefinedRecord {
