+++ src/language/type/inspector/index.ts
