+++ src/language/type/typeUtils/factories.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT license.
 
 import { Type } from "..";
-import { Assert } from "../../common";
-import { ParameterScopeItem } from "../../inspection";
+import { Assert } from "../../../common";
+import { ParameterScopeItem } from "../../../inspection";
 import { PrimitiveTypeConstantMap, primitiveTypeMapKey, typeKindFromPrimitiveTypeConstantKind } from "./primitive";
 import { dedupe } from "./typeUtils";
