+++ src/test/libraryTest/inspection/typeUtils.ts
@@ -4,8 +4,8 @@
 import { expect } from "chai";
 import "mocha";
 import "mocha";
-import { Type, TypeUtils } from "../../../type";
-import { primitiveTypeFactory } from "../../../type/typeUtils";
+import { Type, TypeUtils } from "../../../language/type";
+import { primitiveTypeFactory } from "../../../language/type/typeUtils";
 
 interface AbridgedType {
     readonly kind: Type.TypeKind;
